Move download functions to helpers module
